+++ tom_d4rt/lib/src/runtime_types.dart
@@ -1979,6 +1979,7 @@ class InterpretedInstance implements RuntimeValue {
     throw UndefinedMemberD4rtException(
       "Undefined property '$name' on ${klass.name}.",
       memberName: name,
+      receiver: this,
     );
   }
 
@@ -2551,6 +2552,7 @@ class InterpretedEnumValue implements RuntimeValue /* Add RuntimeValue */ {
     throw UndefinedMemberD4rtException(
       "Undefined property '$memberName' on enum value '$this'.",
       memberName: memberName,
+      receiver: this,
     );
   }